refactor(claude): extract calculateModelCost to keep claude.lib.mjs under 1500 lines
The R1 web_search billing change in the previous commit pushed
src/claude.lib.mjs from 1496 to 1519 lines, tripping the
check-file-line-limits CI job (1500-line cap, see issue #1593).

Move calculateModelCost into a new file src/claude.cost.lib.mjs with
identical behaviour. The function is re-exported from claude.lib.mjs so
callers continue to import it from the same path. claude.lib.mjs is now
back at 1458 lines.

No behaviour change.
